+++ packages/antd-management-fast-common/src/utils/logAssist.js
@@ -1,0 +1,8 @@
+export function logTemplate(
+  o,
+  division = '================================================================================',
+) {
+  console.log(division);
+
+  console.log(o);
+}